Rename Encoders and Decoders
Since we had 2 types (Encoders and Decoders) and a lot of cases,
all of them are now unified and named expressively.

Everything is encoder, and it encodes TypeA to TypeB, so
naming is TypeAToTypeBEncoder. It should be clear from name what does
encoder do and whic types are expected.

Also, namespace is changed, so they all reside inside Encoding namespace