[Bugfix]: Enable building empty relationship data
There was a subtle bug caused by too loose checks performed on resource
identifier collection - if an empty array, representing an empty to-many
relationship, was being converted to an empty relationship collection,
the result would evaluate to an empty to-one relationship instead.

This commit also adds tests to cover this case, as well as tests
other cases and for building empty relationship data objects.